// src/enums.ts
export enum DateFormat {
  ISO = "YYYY-MM-DD",
  US = "MM/DD/YYYY",
  EU = "DD.MM.YYYY",
  VERBOSE = "MMMM DD, YYYY",
}

export enum TimeUnit {
  MILLISECONDS = "milliseconds",
  SECONDS = "seconds",
  MINUTES = "minutes",
  HOURS = "hours",
  DAYS = "days",
  WEEKS = "weeks",
  MONTHS = "months",
  YEARS = "years",
}